Fix problem with cubes without beam header key
Note that the beam size is only used when size filtering. This is typically not done anymore. Either way, the bug should now be fixed. The lines extracting the beam information are now in the 'sizeFilter' part of the code, and the beam is set to zero if no information is present in the header.