Update store_xml.py
Added function to store the full information from the source finding (in addition to the results from the parametrization). Additionally, it now accepts a list of parameters to write if the user does not wish to store all parameters. The old output function can be deleted if the new one is tested successfully.