Removed header information from parametrisation
As parametrisation is meant to be happening in pixel coordinates, the FITS file header is no longer required in the parametrisation module. I removed the header data from all functions, and no more header information is passed on to the parametrisation module. This also resolves the issue that data cubes with excessive headers brought SoFiA to a halt.